Reorganized the namespaces and file locations to be less tied to WindEditor and less compartmentalized.
