+++ ArchiveToolsLib/Archive/RARCReader.cs
@@ -2,9 +2,9 @@
 using System.Collections.Generic;
 using System.IO;
 using System.Text;
-using WEditor.FileSystem;
+using WArchiveTools.FileSystem;
 
-namespace WArchiveTools.rarc
+namespace WArchiveTools.Archive
 {
     public partial class RARC
     {
