+++ ArchiveToolsLib/ArchiveToolsLib.csproj
@@ -41,13 +41,13 @@
     <Compile Include="ArcUtilities.cs" />
     <Compile Include="FileSystem\VirtualFilesystem.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="rarc\RARC.cs" />
-    <Compile Include="rarc\RARCPacker.cs" />
-    <Compile Include="rarc\RARCReader.cs" />
-    <Compile Include="yay0\Yay0Decoder.cs" />
-    <Compile Include="yay0\Yay0Encoder.cs" />
-    <Compile Include="yaz0\Yaz0Decoder.cs" />
-    <Compile Include="yaz0\Yaz0Encoder.cs" />
+    <Compile Include="Archive\RARC.cs" />
+    <Compile Include="Archive\RARCPacker.cs" />
+    <Compile Include="Archive\RARCReader.cs" />
+    <Compile Include="Compression\Yay0Decoder.cs" />
+    <Compile Include="Compression\Yay0Encoder.cs" />
+    <Compile Include="Compression\Yaz0Decoder.cs" />
+    <Compile Include="Compression\Yaz0Encoder.cs" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
