+++ ArchiveToolsLib/Compression/Yaz0Encoder.cs
